feat!: support 5-state logic simulation
Extends the Logic enum to allow evaluation AND, OR, and NOT for
the variants High, Low, D, Dbar, X.

Add tests for logical operations on Logic type,
and for simulating with D and X states.

BREAKING CHANGE: The simulator no longer initializes all
initial net states to Unassigned (X). Instead, the mapping
starts out empty, and any missing keys are "unassigned".
This allows simulating with 'X' inputs